Rename whileSpeaking to whileSpeakIncomingActivity
